Add tests that demonstrate a query bug
It looks like using a pull spec in the same query as a :with clause
results in extra values being added to the results. I suspect the
":with variables" are being left in the results rather than being
stripped out, though the specific way this is happening is still
confusing to me.